sphinx-agent: Move `Uri` type into the sphinx-agent
Decoupling it from `pygls` and using it within the
sphinx-agent *should* allow us to obtain a consistent representation
for files across OSes, Sphinx versions, filepaths vs uris etc!

Hopefully!!
